backend/requirements.txt
requests
python-dotenv 
pymongo
beautifulsoup4
requests
langchain 
langchain-together
